bist-equity-research/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.8.0
plotly>=5.18.0
reportlab>=4.1.0
ta>=0.11.0
pandas-ta>=0.3.14b1
evdspy>=1.1.0
isyatirimhisse>=2.0.0
apify-client>=1.7.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
mplfinance>=0.12.10b0
kaleido>=0.2.1
schedule>=1.2.0
python-dotenv>=1.0.0
